requirements.txt
jishaku
toml~=0.10.1
uvloop~=0.14.0
parsedatetime~=2.6
pytz
discord.py[voice]
tortoise-orm[accel]
aerich
asyncpg
python-rapidjson
ciso8601
Babel~=2.8.0
python-cas
aiohttp[speedups]~=3.6.2
python-dateutil~=2.6.1
aiohttp_session
cryptography